Update patch set 1
Patch Set 1:

> I would prefer to deprecate them and recreate them with better
 > names.

Cool. Let's do it then.
I'll be out for the rest of the week ... but since I do not have
the required permissions ... how would I get this started on Monday?
Officially ask on repo-discuss, or do I just wait for a kind
Admin to do it?

Patch-set: 1